Original check-in
Templates for all languages.
Note that these are currently duplicated from
the Eclipse chameleon plugins, as these projects
each currently need to contain the templates for
each of the specific programming languages.

Over time, this repo should be the single
location for templates, but currently they
are located in the various plugins, such as
Eclipse and Vim.

There are also many automatically generated
templates that are not finished.  They require
some manual changes before being moved/renamed
from *TODO* to a name without the "TODO" in it.